pyproject.toml
[project]
name = "nano-perplexity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "backoff>=2.2.1",
    "beautifulsoup4>=4.12.3",
    "fastapi>=0.115.7",
    "googlesearch-python>=1.3.0",
    "gradio>=5.13.2",
    "json-repair>=0.35.0",
    "lxml>=5.3.0",
    "nicegui>=2.10.1",
    "openai>=1.60.2",
    "pydantic>=2.10.6",
    "requests>=2.32.3",
    "tavily-python>=0.5.0",
    "uvicorn>=0.34.0",
]
